feat: app events

+++ app/routes/api.tickets.$id.tsx
@@ -1,9 +1,12 @@
 import {authenticate} from "../shopify.server";
+import {Ticket} from "../types";
+import {onTicketDeleted, onTicketEdited} from "../app-events.service";
+import type {ActionFunctionArgs, LoaderFunctionArgs} from "react-router";
 
 // eslint-disable-next-line no-undef
 const MOCK_API_URL = process.env.MOCK_API_URL || "";
 
-export const loader = async ({request, params}) => {
+export const loader = async ({request, params}: LoaderFunctionArgs) => {
   await authenticate.admin(request);
 
   const {id} = params;
@@ -12,8 +15,10 @@ export const loader = async ({request, params}) => {
   return {ticket};
 };
 
-export const action = async ({request, params}) => {
+export const action = async ({request, params}: ActionFunctionArgs) => {
   await authenticate.admin(request);
+
+  console.dir(request, {depth: 100});
 
   const {id} = params;
 
@@ -26,14 +31,18 @@ export const action = async ({request, params}) => {
         ...body
       }),
     });
-    const ticket = await res.json();
 
-    console.dir({ticket});
+    const ticket: Ticket = await res.json();
+
+    onTicketEdited(ticket);
+
     return {ticket};
   }
 
-  if (request.method === "DELETE") {
+  if (request.method === "DELETE" && id) {
     await fetch(`${MOCK_API_URL}/${id}`, {method: "DELETE"});
+    onTicketDeleted(id);
+
     return {ok: true};
   }
